feat: add `GetAllUsers` method and update user handling constructors
- Added `GetAllUsers` method in `GetUserService` and `GetAll` in `GetUserQueryHandler` to retrieve all users.
- Enhanced `GetUserViewModel` constructor for nullable numerical values.
- Refactored imports and service assignments in `index.ts`.
