+++ user-service/src/Application/DTO/Query/VerifyIfUsersExistsByUuidsDTO.ts
@@ -1,6 +1,6 @@
 export class VerifyIfUsersExistsByUuidsDTO
 {
-    constructor(public readonly Uuids: string[])
+    constructor(public readonly Uuids: (string | null)[])
     {
     }
 }
